Refactored loadfiles input to be generic
