+++ src/app/features/sidebar/hooks/index.ts
@@ -1,20 +1,7 @@
-import {useDrop} from "react-dnd"
-import {NativeTypes} from "react-dnd-html5-backend"
-import {DragItem, DragProps} from "src/app/features/import/use-import-on-drop"
 import {useBrimApi} from "src/app/core/context"
+import {useFilesDrop} from "src/util/hooks/use-files-drop"
 
 export const useQueryImportOnDrop = () => {
   const api = useBrimApi()
-  return useDrop<DragItem, unknown, DragProps>(() => ({
-    accept: [NativeTypes.FILE],
-    drop: ({files}) => {
-      if (files && files[0]) {
-        api.queries.import(files[0])
-      }
-    },
-    collect: (m) => ({
-      isOver: m.isOver(),
-      canDrop: m.canDrop(),
-    }),
-  }))
+  return useFilesDrop({onDrop: (files) => api.queries.import(files[0])})
 }
